[PT-3] create bloc for cart

+++ domain/lib/models/cart/cart.dart
@@ -2,14 +2,14 @@ import 'cart_dish.dart';
 
 class Cart {
   final List<CartDish> cartDishes;
-  final int totalCost;
+  final double totalCost;
 
   const Cart({
     required this.cartDishes,
     required this.totalCost,
   });
 
-  Cart copyWith({List<CartDish>? cartDishes, int? totalCost}) {
+  Cart copyWith({List<CartDish>? cartDishes, double? totalCost}) {
     return Cart(
       cartDishes: cartDishes ?? this.cartDishes,
       totalCost: totalCost ?? this.totalCost,
